requirements.txt
python-rtmidi
mido
pygame